refactor(paper): treat r_task as 0/1 int, drop _binarize_r_task and >0.5 gates
TB 2.0 task verifiers' ``tests/test.sh`` already write ``0`` or ``1``
to ``/logs/verifier/reward.txt`` — the harbor ``VerifierResult.rewards``
schema is ``float`` only because the framework supports partial-credit
tasks, but our in-tree tasks are already binary at the source. The
``_binarize_r_task(r_task, threshold=0.5)`` helper plus five inline
``r_task > 0.5`` / ``r_task <= 0.5`` / ``r_task != 0.0`` checks
were rounding values that are already 0 or 1 — pure noise.

Changes:
- ``_harbor_r_task`` now returns ``int`` (0 or 1) via
  ``int(round(float(reward)))`` at the source. Docstring documents
  the binary contract; callers use the value directly as a bool.
- ``_binarize_r_task`` deleted (single-call helper).
- ``_q_update_from_subtask`` (r_task:int): removed ``r_task_bin``
  alias; the trace dict's ``"r_task"`` field uses ``r_task`` directly;
  ``if r_task_bin > 0.5`` → ``if r_task``.
- ``_attribution_and_extract_dispatch`` (r_task:int): Rule 2
  ``if r_task > 0.5 and ...`` → ``if r_task and ...``; Rule 5
  ``elif r_task <= 0.5 and ...`` → ``elif not r_task and ...``.
- ``_incremental_edit_on_failure`` (r_task:int): ``if r_task != 0.0 or ...``
  → ``if not r_task or ...``.
- ``NearMissRefiner.is_near_miss`` (r_task:int): ``r_task == 0.0``
  → ``r_task == 0``.

93/93 tests pass. The five ``0.5`` magic numbers and one helper
collapse to a single ``int(round(...))`` at the source. Type system
now documents the binary assumption; if a future partial-credit task
is added, the call site can decide explicitly how to handle it
rather than every consumer silently re-applying the threshold.

+++ skillq/method/near_miss.py
@@ -60,12 +60,12 @@ class NearMissRefiner:
 
     def is_near_miss(
         self,
-        r_task: float,
+        r_task: int,
         q_value: float,
         theta_near_miss: float,
     ) -> bool:
         """Return ``True`` when a failure is a near-miss (Eq. 11)."""
-        return r_task == 0.0 and q_value >= theta_near_miss
+        return r_task == 0 and q_value >= theta_near_miss
 
     def propose_edit(
         self,
